Fix import paths in (moved) claims files
